pyproject.toml
[project]
name = "url-shortener-scalable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "click>=8.2.1",
    "email-validator>=2.3.0",
    "fakeredis>=2.31.1",
    "fastapi[standard]>=0.116.1",
    "flower>=2.0.1",
    "jinja2>=3.1.6",
    "mongomock>=4.3.0",
    "motor>=3.7.1",
    "psycopg2-binary>=2.9.10",
    "pulumi>=3.196.0",
    "pulumi-aws>=7.7.0",
    "pulumi-command>=1.1.0",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.15.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "pytz<2022",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "fakeredis>=2.23.0",
    "mongomock>=4.1.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["create_service*", "common*"]
exclude = ["tests*"]
